+++ packages/frameworks/vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mirrorworld/mirage.vue
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @mirrorworld/mirage.core@1.5.0
+
 ## 0.3.1
 
 ### Patch Changes